@@ -1,10 +1,11 @@
 package com.empodera;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.method.LinkMovementMethod;
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class LearnSubjectActivity extends AppCompatActivity {
 
